+++ tools/check_format_compatible.sh
@@ -58,7 +58,7 @@ generate_db()
 compare_db()
 {
     set +e
-    $script_copy_dir/verify_random_db.sh $1 $2 $3
+    $script_copy_dir/verify_random_db.sh $1 $2 $3 $4
     if [ $? -ne 0 ]; then
         echo ==== Read different content from $1 and $2 or error happened. ====
         exit 1
@@ -95,7 +95,7 @@ generate_db $input_data_path $compare_base_db_dir
 for checkout_obj in "${checkout_objs[@]}"
 do
    echo == Opening DB from "$checkout_obj" using debug build of $checkout_flag ...
-   compare_db $test_dir/$checkout_obj $compare_base_db_dir db_dump.txt
+   compare_db $test_dir/$checkout_obj $compare_base_db_dir db_dump.txt 1
 done
 
 for checkout_obj in "${forward_compatible_checkout_objs[@]}"
@@ -104,7 +104,7 @@ do
    git checkout origin/$checkout_obj
    make clean
    make ldb -j32
-   compare_db $test_dir/$checkout_obj $compare_base_db_dir forward_${checkout_obj}_dump.txt
+   compare_db $test_dir/$checkout_obj $compare_base_db_dir forward_${checkout_obj}_dump.txt 0
 done
 
 echo ==== Compatibility Test PASSED ====