+++ HISTORY.md
@@ -2,6 +2,7 @@
 ## Unreleased
 ### New Features
 * DB::ResetStats() to reset internal stats.
+* ldb add option --try_load_options, which will open DB with its own option file.
 
 ## 5.4.0 (04/11/2017)
 ### Public API Change